+++ projects/swimlane/ngx-ui/src/lib/components/date-range-calendar/models/date-range.model.ts
@@ -10,3 +10,11 @@ export interface PresetConfig {
   expression?: string;
   range?: () => [Date | null, Date | null];
 }
+
+export interface TooltipDateItem {
+  key: string;
+  value: {
+    clip: string;
+    display: string;
+  };
+}