+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## HEAD (unreleased)
 
+- Enhancement (`ngx-filter`): Added new inputs to show tooltip for customDropdown
+- Enhancement (`ngx-date-range-picker`): Added tooltip
+- Enhancement (`ngx-date-range-picker`): Added future preset values
+- Enhancement (`ngx-date-range-picker`): Input Fields for preset values display relative expressions (e.g., now-2d to now) instead of converted date values for better clarity and editing.
+
 ## 50.0.0-alpha.5 (2025-07-24)
 
 - Fix (`ngx-column`): Content should display when column child is active on load.